+++ core-bundle/tests/Contao/StringUtilTest.php
@@ -19,6 +19,7 @@ use Contao\System;
  *
  * @author Yanick Witschi <https://github.com/toflar>
  * @author Martin Auswöger <martin@auswoeger.com>
+ * @author Leo Feyer <https://github.com/leofeyer>
  *
  * @group legacy
  */
@@ -121,6 +122,26 @@ class StringUtilTest extends TestCase
     public function testParseSimpleTokensInvalidComparison($string)
     {
         StringUtil::parseSimpleTokens($string, ['foo' => 'bar']);
+    }
+
+    /**
+     * Tests the standardize() method.
+     */
+    public function testStandardize()
+    {
+        $GLOBALS['TL_CONFIG']['characterSet'] = 'UTF-8';
+
+        $this->assertEquals('foo', StringUtil::standardize('foo'));
+        $this->assertEquals('foo', StringUtil::standardize('FOO'));
+        $this->assertEquals('FOO', StringUtil::standardize('FOO', true));
+        $this->assertEquals('foo-bar', StringUtil::standardize('foo bar'));
+        $this->assertEquals('foo-bar', StringUtil::standardize('%foo&bar~'));
+        $this->assertEquals('foo-bar', StringUtil::standardize('foo&amp;bar'));
+        $this->assertEquals('foo-bar', StringUtil::standardize('foo-{{link::12}}-bar'));
+        $this->assertEquals('foo-bar', StringUtil::standardize('föö-bär'));
+        $this->assertEquals('id-123', StringUtil::standardize('123'));
+        $this->assertEquals('123foo', StringUtil::standardize('123foo'));
+        $this->assertEquals('foo123', StringUtil::standardize('foo123'));
     }
 
     /**